+++ samples/bluetooth/peripheral/prj.conf
@@ -7,3 +7,4 @@ CONFIG_BT_ATT_PREPARE_COUNT=2
 CONFIG_BT_PRIVACY=y
 CONFIG_BT_DEVICE_NAME="Test peripheral"
 CONFIG_BT_DEVICE_APPEARANCE=833
+CONFIG_BT_BAS=y
