+++ samples/bluetooth/peripheral_csc/prj.conf
@@ -4,3 +4,4 @@ CONFIG_BT=y
 CONFIG_BT_PERIPHERAL=y
 CONFIG_BT_DEVICE_NAME="CSC peripheral"
 CONFIG_BT_DEVICE_APPEARANCE=1157
+CONFIG_BT_BAS=y
