+++ samples/bluetooth/peripheral_csc/src/main.c
@@ -20,9 +20,9 @@
 #include <bluetooth/conn.h>
 #include <bluetooth/uuid.h>
 #include <bluetooth/gatt.h>
+#include <bluetooth/bas.h>
 
 #include <gatt/dis.h>
-#include <gatt/bas.h>
 
 #define DEVICE_NAME			CONFIG_BT_DEVICE_NAME
 #define DEVICE_NAME_LEN			(sizeof(DEVICE_NAME) - 1)
