+++ samples/bluetooth/peripheral_esp/prj.conf
@@ -4,3 +4,4 @@ CONFIG_BT_SMP=y
 CONFIG_TINYCRYPT=y
 CONFIG_BT_DEVICE_NAME="ESP peripheral"
 CONFIG_BT_DEVICE_APPEARANCE=768
+CONFIG_BT_BAS=y
