+++ samples/bluetooth/peripheral_hids/prj.conf
@@ -4,3 +4,4 @@ CONFIG_BT_SMP=y
 CONFIG_BT_PERIPHERAL=y
 CONFIG_BT_DEVICE_NAME="Test HoG mouse"
 CONFIG_BT_DEVICE_APPEARANCE=962
+CONFIG_BT_BAS=y
