+++ samples/bluetooth/peripheral_hids/src/Makefile
@@ -1,3 +1,3 @@
 ccflags-y +=-I${ZEPHYR_BASE}/samples/bluetooth
 
-obj-y = main.o dis.o bas.o hog.o
+obj-y = main.o dis.o hog.o
