--- samples/bluetooth/peripheral_hids/src/bas.c
@@ -1,2 +1,0 @@
-/* Workaround build system bug that will put objects in source dir */
-#include "../../gatt/bas.c"
