+++ samples/bluetooth/peripheral_hids/src/main.c
@@ -19,9 +19,9 @@
 #include <bluetooth/conn.h>
 #include <bluetooth/uuid.h>
 #include <bluetooth/gatt.h>
+#include <bluetooth/bas.h>
 
 #include <gatt/dis.h>
-#include <gatt/bas.h>
 #include <gatt/hog.h>
 
 #define DEVICE_NAME		CONFIG_BT_DEVICE_NAME
