+++ samples/bluetooth/peripheral_hr/prj.conf
@@ -4,3 +4,4 @@ CONFIG_BT_SMP=y
 CONFIG_BT_PERIPHERAL=y
 CONFIG_BT_DEVICE_NAME="Zephyr Heartrate Sensor"
 CONFIG_BT_DEVICE_APPEARANCE=833
+CONFIG_BT_BAS=y
