+++ samples/bluetooth/peripheral_hr/src/main.c
@@ -19,10 +19,10 @@
 #include <bluetooth/conn.h>
 #include <bluetooth/uuid.h>
 #include <bluetooth/gatt.h>
+#include <bluetooth/bas.h>
 
 #include <gatt/hrs.h>
 #include <gatt/dis.h>
-#include <gatt/bas.h>
 
 #define DEVICE_NAME		CONFIG_BT_DEVICE_NAME
 #define DEVICE_NAME_LEN		(sizeof(DEVICE_NAME) - 1)
