+++ subsys/bluetooth/Makefile
@@ -1,4 +1,5 @@
 obj-y += common/
+obj-y += lib/
 obj-$(CONFIG_BT_SHELL) += shell/
 obj-$(CONFIG_BT_HCI) += host/
 obj-$(CONFIG_BT_CTLR) += controller/
