+++ subsys/bluetooth/lib/Makefile
@@ -1,0 +1,1 @@
+obj-$(CONFIG_BT_BAS) += bas.o
