+++ subsys/bluetooth/lib/bas.c
